Access nests, events, detections, and totals

Access nests, events, detections, abundance, and density from simulation objects.

get_nests(x, ...)
# S3 method for bsims_population
get_nests(x, ...)

get_events(x, ...)
# S3 method for bsims_events
get_events(x, ...)

get_detections(x, ...)
# S3 method for bsims_detections
get_detections(x, ...)

get_abundance(x, ...)
# S3 method for bsims_population
get_abundance(x, ...)

get_density(x, ...)
# S3 method for bsims_population
get_density(x, ...)

Arguments

x
simulation object.

...
other arguments passed to internal functions.

Details

get_nests extracts the next locations.

get_events extracts the events.

get_detections extracts the detections.

get_abundance gets total abundance (N), get_density gets average density (abundance/area: N/A).

Value

get_abundance and get_density returns a non-negative numeric value.

get_nests returns a data frame with the following columns: i individual identifier, s spatial stratum (H: habitat, E: edge, R: road) x and y are coordinates of the nest locations, g is behavioural (mixture) group or NA.

get_events returns a data frame with the following columns: x and y are locations of the individual at the time of the event, t time of the event within the duration interval, v indicator variable for vocal (1) vs. movement (0) event, i individual identifier.

get_detections returns a data frame with the following columns: x and y are locations of the individual at the time of the event, t time of the event within the duration interval, v indicator variable for vocal (1) vs. movement (0) event, d distance from observer when detected (otherwise NA). i individual identifier, j perceived individual identifier.

See also

Examples

phi <- 0.5                 # singing rate
tau <- 1:3                 # EDR by strata
dur <- 10                  # simulation duration
tbr <- c(3, 5, 10)         # time intervals
rbr <- c(0.5, 1, 1.5, Inf) # counting radii

l <- bsims_init(10, 0.5, 1)# landscape
p <- bsims_populate(l, 1)  # population
e <- bsims_animate(p,      # events
  vocal_rate=phi, duration=dur)
d <- bsims_detect(e,       # detections
  tau=tau)
x <- bsims_transcribe(d,   # transcription
  tint=tbr, rint=rbr)

## next locations
head(get_nests(p))

#>   i s         x         y  g
#> 1 1 H -2.742498  1.334694 NA
#> 2 2 H -2.204781 -3.317068 NA
#> 3 3 H -3.836963  2.301717 NA
#> 4 4 H -2.824854 -1.563529 NA
#> 5 5 H -2.920065  1.645729 NA
#> 6 6 H -3.677161 -2.103441 NA

head(get_nests(e))

#>   i s         x         y  g
#> 1 1 H -2.742498  1.334694 G1
#> 2 2 H -2.204781 -3.317068 G1
#> 3 3 H -3.836963  2.301717 G1
#> 4 4 H -2.824854 -1.563529 G1
#> 5 5 H -2.920065  1.645729 G1
#> 6 6 H -3.677161 -2.103441 G1

head(get_nests(d))

#>   i s         x         y  g
#> 1 1 H -2.742498  1.334694 G1
#> 2 2 H -2.204781 -3.317068 G1
#> 3 3 H -3.836963  2.301717 G1
#> 4 4 H -2.824854 -1.563529 G1
#> 5 5 H -2.920065  1.645729 G1
#> 6 6 H -3.677161 -2.103441 G1

head(get_nests(x))

#>   i s         x         y  g
#> 1 1 H -2.742498  1.334694 G1
#> 2 2 H -2.204781 -3.317068 G1
#> 3 3 H -3.836963  2.301717 G1
#> 4 4 H -2.824854 -1.563529 G1
#> 5 5 H -2.920065  1.645729 G1
#> 6 6 H -3.677161 -2.103441 G1

## abundance
get_abundance(p)

#> [1] 89

get_abundance(e)

#> [1] 89

get_abundance(d)

#> [1] 89

get_abundance(x)

#> [1] 89

## density
get_density(p)

#> [1] 0.89

get_density(e)

#> [1] 0.89

get_density(d)

#> [1] 0.89

get_density(x)

#> [1] 0.89

## events
head(get_events(e))

#>            x           y          t v  i
#> 1  0.9787458  0.02728511 0.02829169 1 56
#> 2 -1.7141816  3.59273318 0.03070180 1 29
#> 3  2.6589502 -2.81234291 0.03541327 1 62
#> 4  3.1968096  2.22174190 0.04634881 1 59
#> 5  0.1004428 -1.51986307 0.05692312 1 40
#> 6 -0.9479958  3.95741887 0.06460994 1 36

head(get_events(d))

#>            x           y          t v  i
#> 1  0.9787458  0.02728511 0.02829169 1 56
#> 2 -1.7141816  3.59273318 0.03070180 1 29
#> 3  2.6589502 -2.81234291 0.03541327 1 62
#> 4  3.1968096  2.22174190 0.04634881 1 59
#> 5  0.1004428 -1.51986307 0.05692312 1 40
#> 6 -0.9479958  3.95741887 0.06460994 1 36

head(get_events(x))

#>            x           y          t v  i
#> 1  0.9787458  0.02728511 0.02829169 1 56
#> 2 -1.7141816  3.59273318 0.03070180 1 29
#> 3  2.6589502 -2.81234291 0.03541327 1 62
#> 4  3.1968096  2.22174190 0.04634881 1 59
#> 5  0.1004428 -1.51986307 0.05692312 1 40
#> 6 -0.9479958  3.95741887 0.06460994 1 36

## detections
head(get_detections(d))

#>             x           y          t v         d  i  j
#> 1   0.9787458  0.02728511 0.02829169 1 0.9791261 56 56
#> 7   1.9624480 -0.12713910 0.10098929 1 1.9665621 84 84
#> 8   0.9383907  0.92551038 0.12573858 1 1.3180086 55 55
#> 9   0.7952968  3.55519005 0.13021546 1 3.6430582 57 57
#> 12 -0.1734305 -0.11315131 0.17962572 1 0.2070781 42 42
#> 15  2.4378719  0.98378620 0.22931232 1 2.6288885 74 74

head(get_detections(x))

#>             x           y          t v         d  i  j
#> 1   0.9787458  0.02728511 0.02829169 1 0.9791261 56 56
#> 7   1.9624480 -0.12713910 0.10098929 1 1.9665621 84 84
#> 8   0.9383907  0.92551038 0.12573858 1 1.3180086 55 55
#> 9   0.7952968  3.55519005 0.13021546 1 3.6430582 57 57
#> 12 -0.1734305 -0.11315131 0.17962572 1 0.2070781 42 42
#> 15  2.4378719  0.98378620 0.22931232 1 2.6288885 74 74